Workaround: download source code from Internet Archive snapshot
The the LipForge page this script tries to download the CRlibm source from has vanished, making it impossible to install the package. Fortunately, the Internet Archive has a snapshot of the source code; I've pointed the download link there as a *temporary* workaround.

Although there may be licensing subtleties I'm not aware of, it seems to me that CRlibm's GPL 2.0 license allows the Internet Archive, like anyone else, to distribute copies of the source code.